Initial import of open source release
This library is the work of several IBMers:

Ian Munsie
Michael Neuling
Philippe Bergheaud
Alistair Popple

A few files that were primarily authored by someone other than myself
will be included in a separate commit.

Test cases are not currently included. These are dependent on specific
AFUs to function and may be released in future along with the AFUs they
depend on.