Use the new ioctl to construct sysfs path to the afu
This patch changes the cxl_sysfs_fd function to use new ioctl
CXL_IOCTL_GET_AFU_ID available on afu file descriptor to fetch the
cxl_afu_id structure containing the card/offset & mode information. This
info is then used to construct the path to the afu's sysfs
directory. Older implementation was dependent on the major/minor number
of the afu retrieved by fstat on the fd. However with new cxl kernel api
its possible for a kernel module to have its own device that exports
libcxl api to the userspace. This broke the existing implementation
which was dependent on the major/minor numbers fetched from the device
file.

If the ioctl is not available (in case of older cxl driver) the
implementation falls back to the previous mechanism of constructing the
sysfs path using the major/minor numbers retrieved by fstating the file
descriptor to the afu device.