Check that cxl.h is recent and download if necessary
We already have code that downloads the cxl.h if it can't be found, but
that isn't enough as the build box may have a distributed cxl.h that is
out of date.

This modifies the test to also check if struct cxl_afu_id is defined in
the header and download an updated version if necessary.

The rule no longer targets the precise header file since we don't want
Make to decide that it is up to date and not run the rule, as we now
implement our own functionality to decide that.

Switches to using curl instead of wget since curl doesn't require hacks
to overwrite destination files.