Allow cxl_afu_fd_to_h() to work with anonymous file descriptors
The new cxl kernel api allows afu file descriptors be be handed out
through afu specific drivers instead of the classic afu user api. These
file descriptors implement the same user api, but since they are not
opened via a character device some of the safety checks we do will fail.

This patch removes the fail condition if the file descriptor is not a
character device, and won't attempt to find the device name from the
major and minor numbers in this case.

If the file is not an afu file descriptor than we will simply fail
a little later when CXL_IOCTL_GET_AFU_ID fails.