Create LIBSONAME link
For the packaging, we should not only include the shared library itself, but
also the symbolic link which ldconfig would produce. (This is necessary, so
that the link gets removed by dpkg, for ex., automatically when the package gets
removed.) And the packaging doesn't have to deal with the soname version, so
putting this in the Makefile while waiting.
Also updating the all and clean targets.