libcxl: Check afu link when read from PSA mmio return all FFs
Presently libcxl will return an error or raise a SIGBUS when it reads
all FFs from PSA mmio space. This prevents returning wrong values to
the caller when the CXL card has fenced due to an eeh error. But this
posses problems as for some AFUs all FFs might be a valid value for a
PSA register and when read libcxl might incorrectly assume the card to
be fenced and indicate an error to the caller.

We fix this problem by memory mapping the afu descriptor from sysfs to
the process address. When libcxl mmio space read functions
cxl_mmio_read{64,32}() read all FFs we check the first word of the
mapped afu descriptor to see if it all FFs. If yes we assume the card
has fenced and we return an error to the caller. Else we assume the
read value as valid and return a success to the caller with all FFs as
result.

This patch introduces three new internal functions:

* cxl_open_afudesc(): Maps the afu sysfs attr named 'afu_desc' to
  process address space.

* cxl_close_afudesc(): Unmaps and closes the afu descriptor from the
  process address space.

* cxl_afu_link_ok(): Reads the first word from mapped afu descriptor
  and returns 1 if its *not* all FFs else returns 0 indicating link
  down. In case there was an error mapping afu_desc it returns '0'

Signed-off-by: Vaibhav Jain <vaibhav@linux.vnet.ibm.com>